feat(#10): masked view on iOS
