+++ apple/common/RNGMaskedViewManager.m
@@ -1,0 +1,5 @@
+#import <React/RCTViewManager.h>
+
+@interface RCT_EXTERN_MODULE(RNGMaskedViewManager, RCTViewManager)
+
+@end
