+++ apple/common/RNGRadialGradientViewManager.swift
@@ -1,5 +1,8 @@
 @objc(RNGRadialGradientViewManager)
 class RNGRadialGradientViewManager: RCTViewManager {
+    @objc override static func requiresMainQueueSetup() -> Bool {
+        return false
+    }
     override func view() -> RCTView? {
         #if os(OSX)
             return RNGRadialGradientView(frame: CGRect(origin: .zero, size: CGSize(width: 0, height: 0)))
