+++ apple/ios/RNGBlurViewManager.swift
@@ -1,5 +1,8 @@
 @objc(RNGBlurViewManager)
 class RNGBlurViewManager: RCTViewManager {
+    @objc override static func requiresMainQueueSetup() -> Bool {
+        return false
+    }
     override func view() -> RCTView? {
         return RNGBlurView()
     }
