+++ src/index.tsx
@@ -1,5 +1,6 @@
 export * from './AngularGradientView';
 export * from './BlurView';
 export * from './LinearGradientView';
+export * from './MaskedView';
 export * from './RadialGradientView';
 export * from './types';
